transferring chat styles to separate files and some layout changes
