+++ src/components/chat/chat-room/chat-room/styles.js
@@ -1,0 +1,11 @@
+import { makeStyles } from '@material-ui/core';
+
+const useStyles = makeStyles(() => ({
+  root: {
+    position: 'absolute',
+    top: '50px',
+    width: '100%',
+  },
+}));
+
+export default useStyles;
