+++ src/components/chat/chat.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { useSelector } from 'react-redux';
 
-import RoomBox from './room-list/room-box';
-import ChatRoom from './chat-room/chat-room';
+import RoomBox from './room-list/room-box/room-box';
+import ChatRoom from './chat-room/chat-room/chat-room';
 
 const Chat = () => {
   const { isVisibleChat } = useSelector(state => state.chat);
