+++ src/components/chat/room-list/header-room-list/header-room-list.jsx
@@ -1,30 +1,20 @@
-import React from 'react';
+import React, { useCallback } from 'react';
 import { useDispatch } from 'react-redux';
 
 import CloseIcon from '@material-ui/icons/Close';
 import IconButton from '@material-ui/core/IconButton';
-import { makeStyles } from '@material-ui/core/styles';
-import * as action from '../../../redux/actions/chat-action';
-import { ChatSvg } from '../../../shared/svgs/menu-items';
+import { hideChat } from '../../../../redux/actions/chat-action';
+import { ChatSvg } from '../../../../shared/svgs/menu-items';
+import useStyles from './styles';
 
-const useStyles = makeStyles(() => ({
-  root: {
-    position: 'absolute',
-    right: 0,
-    color: 'grey',
-  },
-  svgCloseIcon: {
-    fontSize: '2.5rem',
-  },
-}));
 
 const HeaderRoomList = () => {
   const classes = useStyles();
   const dispatch = useDispatch();
 
-  const handleClickHideChat = () => {
-    dispatch(action.hideChat());
-  };
+  const handleClickHideChat = useCallback(() => {
+    dispatch(hideChat());
+  }, []);
 
   return (
     <div className="room-list-header">
