+++ src/components/chat/room-list/header-room-list/styles.js
@@ -1,0 +1,14 @@
+import { makeStyles } from '@material-ui/core/styles';
+
+const useStyles = makeStyles(() => ({
+  root: {
+    position: 'absolute',
+    right: 0,
+    color: 'grey',
+  },
+  svgCloseIcon: {
+    fontSize: '2.5rem',
+  },
+}));
+
+export default useStyles;
