+++ src/components/chat/room-list/room-box/styles.js
@@ -1,0 +1,12 @@
+import { makeStyles } from '@material-ui/core';
+
+const useStyles = makeStyles(() => ({
+  spinner: {
+    position: 'absolute',
+    left: '50%',
+    top: '50%',
+    transform: 'translate(-50%,-50%)',
+  },
+}));
+
+export default useStyles;
