+++ src/components/chat/room-list/room-list/styles.js
@@ -1,0 +1,30 @@
+import { makeStyles } from '@material-ui/core';
+
+const useStyles = makeStyles(theme => ({
+  root: {
+    height: '100%',
+    overflow: 'auto',
+    '& span': {
+      fontWeight: 'bold',
+    },
+    [theme.breakpoints.down('xs')]: {
+      height: 'calc(100vh - 92px);',
+      '& span': {
+        fontSize: '16px !important',
+      },
+    },
+  },
+  avatar: {
+    background: '#43D4DD',
+    border: '2px solid #e83e8c',
+  },
+  dots: {
+    color: '#43D4DD',
+  },
+  customBadge: {
+    backgroundColor: '#76dc37',
+    color: 'white',
+  },
+}));
+
+export default useStyles;
