+++ src/components/chat/room-list/search-room-list/styles.js
@@ -1,0 +1,22 @@
+import { makeStyles } from '@material-ui/core/styles';
+
+const useStyles = makeStyles(theme => ({
+  root: {
+    padding: '2px 4px',
+    display: 'flex',
+    alignItems: 'center',
+    height: '32px',
+    width: '82%',
+    border: '1px solid grey',
+    marginLeft: theme.spacing(1),
+  },
+  input: {
+    marginLeft: theme.spacing(1),
+    flex: 1,
+  },
+  iconButton: {
+    padding: 10,
+  },
+}));
+
+export default useStyles;
