+++ src/redux/actions/chat-action.js
@@ -1,6 +1,22 @@
 import firebase from 'firebase';
-import * as actionTypes from '../constants/chat-types';
-import { fetchJoinedTribeChats, fetchOwnTribeChat, fetchPrivateChats, getTribeInfoAsync } from '../../shared/lib/chat';
+import {
+  SET_ROOM_LIST,
+  CLEAR_CURRENT_ROOM_NAME,
+  CLEAR_ID_ROOM,
+  SHOW_CHAT,
+  HIDE_CHAT,
+  CLEAR_MESSAGES,
+  INCREASE_LIMIT,
+  SET_INITIAL_LIMIT,
+  FETCH_MESSAGES_START,
+  FETCH_MESSAGES_SUCCESS,
+  SET_ROOM_INFO,
+  FETCH_ROOMS_START,
+  CLEAR_ROOMS,
+  FETCH_ROOMS_SUCCESS,
+  FETCH_ROOMS_FAIL,
+  UPDATE_CHATS,
+} from '../constants/chat-types';
 import { getCookies } from '../../shared/lib/authentication';
 
 const snapshotToArray = snapshot => {
@@ -15,58 +31,38 @@ const snapshotToArray = snapshot => {
   return returnArr;
 };
 
+const getTribeInfo = tribeCode =>
+  firebase
+    .database()
+    .ref(`Tribes/${tribeCode}`)
+    .once('value', snapshot => snapshot);
+
+const getTribeInfoAsync = tribeCode =>
+  new Promise(res => {
+    firebase
+      .database()
+      .ref(`Tribes/${tribeCode}`)
+      .once('value', snapshot => res(snapshot.val()));
+  });
+
 export const setRoomLIst = () => ({
-  type: actionTypes.SET_ROOM_LIST,
+  type: SET_ROOM_LIST,
   chatStatus: 'roomlist',
 });
 
-export const setRoomInfo = (idRoom, currentRoomName) => ({
-  type: actionTypes.SET_ROOM_INFO,
-  chatStatus: 'chatroom',
-  idRoom,
-  currentRoomName,
-});
-
-export const clearCurrentRoomName = () => ({
-  type: actionTypes.CLEAR_CURRENT_ROOM_NAME,
-});
-
-export const clearIdRoom = () => ({
-  type: actionTypes.CLEAR_ID_ROOM,
-});
-
-export const showChat = () => ({
-  type: actionTypes.SHOW_CHAT,
-});
-
-export const hideChat = () => ({
-  type: actionTypes.HIDE_CHAT,
-});
-
-export const clearMessages = () => ({
-  type: actionTypes.CLEAR_MESSAGES,
-});
-
-export const increaseLimit = () => ({
-  type: actionTypes.INCREASE_LIMIT,
-});
-
-export const setInitialLimit = () => ({
-  type: actionTypes.SET_INITIAL_LIMIT,
-});
-
-const fetchMessagesStart = () => ({
-  type: actionTypes.FETCH_MESSAGES_START,
-});
-
-export const fetchMessagesSuccess = messages => ({
-  type: actionTypes.FETCH_MESSAGES_SUCCESS,
-  payload: messages,
-});
+export const clearCurrentRoomName = () => ({ type: CLEAR_CURRENT_ROOM_NAME });
+export const clearIdRoom = () => ({ type: CLEAR_ID_ROOM });
+export const showChat = () => ({ type: SHOW_CHAT });
+export const hideChat = () => ({ type: HIDE_CHAT });
+export const clearMessages = () => ({ type: CLEAR_MESSAGES });
+export const increaseLimit = () => ({ type: INCREASE_LIMIT });
+export const setInitialLimit = () => ({ type: SET_INITIAL_LIMIT });
 
 export const fetchMessages = () => (dispatch, getState) => {
   const { idRoom, limit } = getState().chat;
-  dispatch(fetchMessagesStart());
+  dispatch({
+    type: FETCH_MESSAGES_START,
+  });
   firebase
     .database()
     .ref('messages/')
@@ -75,11 +71,10 @@ export const fetchMessages = () => (dispatch, getState) => {
     .equalTo(idRoom)
     .on('value', snapshot => {
       const messages = snapshotToArray(snapshot);
-      if (messages.length !== 0) {
-        dispatch(fetchMessagesSuccess(messages));
-      } else {
-        dispatch(fetchMessagesSuccess(null));
-      }
+      dispatch({
+        type: FETCH_MESSAGES_SUCCESS,
+        payload: messages,
+      });
     });
 };
 
@@ -104,8 +99,7 @@ const sendJoinMessage = (dispatch, firstname, idRoom, roomname) => {
   newMessage.set(chat);
 };
 
-export const enterChatRoom = (idRoom, name) => {
-  const roomname = name;
+export const enterChatRoom = (idRoom, roomname) => {
   const { userFirstName } = getCookies();
   return dispatch => {
     firebase
@@ -117,7 +111,6 @@ export const enterChatRoom = (idRoom, name) => {
         let roomuser = [];
         roomuser = snapshotToArray(snapshot);
         const user = roomuser.find(x => x.firstname === userFirstName);
-
         if (user) {
           const userRef = firebase.database().ref(`roomusers/${user.key}`);
           userRef.update({ status: 'online' });
@@ -137,24 +130,20 @@ export const enterChatRoom = (idRoom, name) => {
           newRoomUser.set(newroomuser);
         }
       });
-
-    dispatch(setRoomInfo(idRoom, roomname));
+    dispatch({
+      type: SET_ROOM_INFO,
+      chatStatus: 'chatroom',
+      idRoom,
+      currentRoomName: roomname,
+    });
   };
 };
 
-const updatePrivateChats = rooms => ({
-  type: actionTypes.UPDATE_PRIVATE_CHATS,
-  rooms,
-});
-
-const updateJoinedTribeChats = rooms => ({
-  type: actionTypes.UPDATE_JOINED_TRIBE_CHATS,
-  rooms,
-});
-
-export const updatePrivateChatsOn = () => {
+const fetchPrivateChats = dispatch => {
   const { userId } = getCookies();
-  return dispatch => {
+  let isInitialFetch = true;
+
+  return new Promise(resolve => {
     firebase
       .database()
       .ref(`Users/${userId}/friends`)
@@ -172,17 +161,78 @@ export const updatePrivateChatsOn = () => {
               unreadMessages: 0,
             });
           });
-          dispatch(updatePrivateChats(privateChats));
+          if (isInitialFetch) {
+            isInitialFetch = false;
+            resolve(privateChats);
+          } else {
+            dispatch({
+              type: UPDATE_CHATS,
+              rooms: privateChats,
+            });
+          }
         } else {
-          dispatch(updatePrivateChats([]));
+          if (isInitialFetch) {
+            isInitialFetch = false;
+            resolve([]);
+          } else {
+            dispatch({
+              type: UPDATE_CHATS,
+              rooms: [],
+            });
+          }
         }
       });
-  };
+  });
 };
 
-export const updateJoinedTribeChatsOn = () => {
+const fetchOwnTribeChat = dispatch => {
   const { userId } = getCookies();
-  return dispatch => {
+  let isInitialFetch = true;
+
+  return new Promise(resolve => {
+    firebase
+      .database()
+      .ref(`Users/${userId}/tribe_code`)
+      .on('value', async snapshot => {
+        const userTribe = snapshot.val();
+        const tribesChats = [];
+        if (userTribe) {
+          const tribe = (await getTribeInfo(userTribe)).val();
+          tribesChats.push({
+            name: tribe.name,
+            idRoom: tribe.code,
+            isPrivateRoom: false,
+            unreadMessages: 0,
+          });
+          if (isInitialFetch) {
+            resolve(tribesChats);
+            isInitialFetch = false;
+          } else {
+            dispatch({
+              type: UPDATE_CHATS,
+              rooms: tribesChats,
+            });
+          }
+        } else {
+          if (isInitialFetch) {
+            isInitialFetch = false;
+            resolve([]);
+          } else {
+            dispatch({
+              type: UPDATE_CHATS,
+              rooms: [],
+            });
+          }
+        }
+      });
+  });
+};
+
+const fetchJoinedTribeChats = dispatch => {
+  const { userId } = getCookies();
+  let isInitialFetch = true;
+
+  return new Promise(resolve => {
     firebase
       .database()
       .ref(`Users/${userId}/tribe_joined`)
@@ -201,44 +251,52 @@ export const updateJoinedTribeChatsOn = () => {
                   unreadMessages: 0,
                 });
               });
-              dispatch(updateJoinedTribeChats(tribesJoinedChats));
+              if (isInitialFetch) {
+                resolve(tribesJoinedChats);
+                isInitialFetch = false;
+              } else {
+                dispatch({
+                  type: UPDATE_CHATS,
+                  rooms: tribesJoinedChats,
+                });
+              }
             })
             .catch(error => {
               console.error('Failed get tribe info', error.message);
             });
         } else {
-          dispatch(updateJoinedTribeChats([]));
+          if (isInitialFetch) {
+            resolve([]);
+            isInitialFetch = false;
+          } else {
+            dispatch({
+              type: UPDATE_CHATS,
+              rooms: [],
+            });
+          }
         }
       });
-  };
+  });
 };
 
-const fetchRoomsStart = () => ({
-  type: actionTypes.FETCH_ROOMS_START,
-});
-
-const fetchRoomsSuccess = rooms => ({
-  type: actionTypes.FETCH_ROOMS_SUCCESS,
-  rooms,
-});
-
-const fetchRoomsFail = () => ({
-  type: actionTypes.FETCH_ROOMS_FAIL,
-});
-
-const clearRooms = () => ({
-  type: actionTypes.CLEAR_ROOMS,
-});
-
 export const fetchRooms = () => dispatch => {
-  dispatch(fetchRoomsStart());
-  dispatch(clearRooms());
-  Promise.all([fetchPrivateChats(), fetchOwnTribeChat(), fetchJoinedTribeChats()])
+  dispatch({
+    type: FETCH_ROOMS_START,
+  });
+  dispatch({
+    type: CLEAR_ROOMS,
+  });
+  Promise.all([fetchPrivateChats(dispatch), fetchOwnTribeChat(dispatch), fetchJoinedTribeChats(dispatch)])
     .then(rooms => {
       const flattedRooms = rooms.flat();
-      dispatch(fetchRoomsSuccess(flattedRooms));
+      dispatch({
+        type: FETCH_ROOMS_SUCCESS,
+        rooms: flattedRooms,
+      });
     })
-    .catch(err => {
-      dispatch(fetchRoomsFail(err.message));
+    .catch(() => {
+      dispatch({
+        type: FETCH_ROOMS_FAIL,
+      });
     });
 };
