+++ src/redux/constants/chat-types.js
@@ -7,12 +7,10 @@ export const HIDE_CHAT = 'HIDE_CHAT';
 export const CLEAR_MESSAGES = 'CLEAR_MESSAGES';
 export const FETCH_MESSAGES_START = 'FETCH_MESSAGES_START';
 export const FETCH_MESSAGES_SUCCESS = 'FETCH_MESSAGES_SUCCESS';
-export const FETCH_MESSAGES_FAIL = 'FETCH_MESSAGES_FAIL';
 export const FETCH_ROOMS_START = 'FETCH_ROOMS_START';
 export const FETCH_ROOMS_SUCCESS = 'FETCH_ROOMS_SUCCESS';
 export const FETCH_ROOMS_FAIL = 'FETCH_ROOMS_FAIL';
 export const CLEAR_ROOMS = 'CLEAR_ROOMS';
 export const INCREASE_LIMIT = 'INCREASE_LIMIT';
 export const SET_INITIAL_LIMIT = 'SET_INITIAL_LIMIT';
-export const UPDATE_PRIVATE_CHATS = 'UPDATE_PRIVATE_CHATS';
-export const UPDATE_JOINED_TRIBE_CHATS = 'UPDATE_JOINED_TRIBE_CHATS';
+export const UPDATE_CHATS = 'UPDATE_CHATS';
