+++ src/redux/reducers/chat-reducer.js
@@ -1,5 +1,21 @@
 import _ from 'lodash';
-import * as actionTypes from '../constants/chat-types';
+import {
+  SET_ROOM_LIST,
+  SET_ROOM_INFO,
+  CLEAR_CURRENT_ROOM_NAME,
+  CLEAR_ID_ROOM,
+  SHOW_CHAT,
+  HIDE_CHAT,
+  CLEAR_MESSAGES,
+  INCREASE_LIMIT,
+  SET_INITIAL_LIMIT,
+  FETCH_MESSAGES_START,
+  FETCH_MESSAGES_SUCCESS,
+  UPDATE_CHATS,
+  CLEAR_ROOMS,
+  FETCH_ROOMS_SUCCESS,
+  FETCH_ROOMS_FAIL,
+} from '../constants/chat-types';
 
 const filterChats = (state, action) => {
   const roomsJoined = [...state.rooms, ...action.rooms];
@@ -16,90 +32,92 @@ const initialState = {
   loadingRooms: true,
   rooms: [],
   limit: 20,
+  errorLoadingRooms: false,
 };
 
 const chatReducer = (state = initialState, action) => {
   switch (action.type) {
-    case actionTypes.SET_ROOM_LIST:
+    case SET_ROOM_LIST:
       return {
         ...state,
         chatStatus: action.chatStatus,
       };
-    case actionTypes.SET_ROOM_INFO:
+    case SET_ROOM_INFO:
       return {
         ...state,
         chatStatus: action.chatStatus,
         idRoom: action.idRoom,
         currentRoomName: action.currentRoomName,
       };
-    case actionTypes.CLEAR_CURRENT_ROOM_NAME:
+    case CLEAR_CURRENT_ROOM_NAME:
       return {
         ...state,
         currentRoomName: '',
       };
-    case actionTypes.CLEAR_ID_ROOM:
+    case CLEAR_ID_ROOM:
       return {
         ...state,
         idRoom: '',
       };
-    case actionTypes.SHOW_CHAT:
+    case SHOW_CHAT:
       return {
         ...state,
         isVisibleChat: true,
       };
-    case actionTypes.HIDE_CHAT:
+    case HIDE_CHAT:
       return {
         ...state,
         isVisibleChat: false,
       };
-    case actionTypes.CLEAR_MESSAGES:
+    case CLEAR_MESSAGES:
       return {
         ...state,
-        messages: null,
+        messages: [],
       };
-    case actionTypes.INCREASE_LIMIT:
+    case INCREASE_LIMIT:
       return {
         ...state,
         limit: state.limit + 5,
       };
-    case actionTypes.SET_INITIAL_LIMIT:
+    case SET_INITIAL_LIMIT:
       return {
         ...state,
         limit: 20,
       };
-    case actionTypes.FETCH_MESSAGES_START:
+    case FETCH_MESSAGES_START:
       return {
         ...state,
         loadingMessages: true,
       };
-    case actionTypes.FETCH_MESSAGES_SUCCESS:
+    case FETCH_MESSAGES_SUCCESS:
       return {
         ...state,
         loadingMessages: false,
         messages: action.payload,
       };
-    case actionTypes.UPDATE_PRIVATE_CHATS:
-      return {
-        ...state,
-        rooms: filterChats(state, action),
-      };
-    case actionTypes.UPDATE_JOINED_TRIBE_CHATS: {
+    case UPDATE_CHATS: {
       return {
         ...state,
         rooms: filterChats(state, action),
       };
     }
-    case actionTypes.CLEAR_ROOMS: {
+    case CLEAR_ROOMS: {
       return {
         ...state,
         rooms: [],
       };
     }
-    case actionTypes.FETCH_ROOMS_SUCCESS: {
+    case FETCH_ROOMS_SUCCESS: {
       return {
         ...state,
         rooms: filterChats(state, action),
         loadingRooms: false,
+      };
+    }
+    case FETCH_ROOMS_FAIL: {
+      return {
+        ...state,
+        errorLoadingRooms: true,
       };
     }
     default:
